requirements.txt
tensorflow_gpu==1.15.0
matplotlib==3.3.1
easydict==1.9
numpy==1.19.1
Cython==0.29.21
scipy==1.1.0
PyYAML==5.3.1
